+++ MIDP/loaders/nifti_loader.py
@@ -102,7 +102,7 @@ class NIfTILoader:
         assert isinstance(prediction, np.ndarray), type(prediction)
         os.makedirs(output_dir, exist_ok=True)
         nib.save(
-            nib.Nifti1Image(prediction.astype(np.int16), affine=np.eye(4)),
+            nib.Nifti1Image(prediction.astype(np.uint8), affine=np.eye(4)),
             os.path.join(output_dir, data_idx + '.nii.gz')
         )
